requirements.txt
astroid==2.5.6
autopep8==1.5.7
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
console-menu==0.6.0
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
progress==1.5
pycodestyle==2.7.0
pylint==2.8.2
python-dotenv==0.17.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
urllib3==1.26.4
wrapt==1.12.1
